@@ -1,6 +1,6 @@
 module coin_chase 
 
-go 1.22.0
+go 1.23.0
 
 require (
 	github.com/google/uuid v1.6.0
