@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.25.2
+// 	protoc-gen-go v1.34.2
+// 	protoc        v5.28.0
 // source: protodef/status.proto
 
 package protodef
@@ -160,7 +160,7 @@ func file_protodef_status_proto_rawDescGZIP() []byte {
 }
 
 var file_protodef_status_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_protodef_status_proto_goTypes = []interface{}{
+var file_protodef_status_proto_goTypes = []any{
 	(*Position)(nil), // 0: status.Position
 	(*Status)(nil),   // 1: status.Status
 }
@@ -179,7 +179,7 @@ func file_protodef_status_proto_init() {
 		return
 	}
 	if !protoimpl.UnsafeEnabled {
-		file_protodef_status_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
+		file_protodef_status_proto_msgTypes[0].Exporter = func(v any, i int) any {
 			switch v := v.(*Position); i {
 			case 0:
 				return &v.state
@@ -191,7 +191,7 @@ func file_protodef_status_proto_init() {
 				return nil
 			}
 		}
-		file_protodef_status_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
+		file_protodef_status_proto_msgTypes[1].Exporter = func(v any, i int) any {
 			switch v := v.(*Status); i {
 			case 0:
 				return &v.state
