@@ -3,7 +3,11 @@ package server_test
 import (
 	"coin_chase/game"
 	"coin_chase/http_server"
-	"strings"
+	"coin_chase/protodef"
+	"context"
+	"log"
+	"sync"
+	"time"
 
 	"encoding/json"
 	"fmt"
@@ -14,6 +18,8 @@ import (
 	"os"
 	"strconv"
 	"testing"
+
+	"google.golang.org/protobuf/proto"
 )
 
 // run test with "go test ./server_test -v -race"
@@ -30,6 +36,7 @@ import (
 type TestClient struct {
 	ID         string
 	Listener   *net.TCPListener
+	Conn       *net.TCPConn
 	WorkerPort int
 }
 
@@ -106,11 +113,7 @@ func TestWorkerPullOut(t *testing.T) {
 				t.Fatalf("failed to create TCP listener: %s", err)
 			}
 
-			clientPort, err := strconv.Atoi(strings.Split(listener.Addr().String(), ":")[1])
-
-			if err != nil {
-				t.Fatalf("invalid port: %s", err)
-			}
+			clientPort := listener.Addr().(*net.TCPAddr).Port
 
 			// 로그인 성공 검사
 			// 유저 아이디 생성
@@ -141,6 +144,17 @@ func TestWorkerPullOut(t *testing.T) {
 				ID:         userId,
 			})
 		}
+
+		for _, cl := range clientListeners {
+			// 이 시점에서 이미 worker들이 dial중이기 때문에 accept가 곧장 진행된다.
+			conn, err := cl.Listener.AcceptTCP()
+
+			if err != nil {
+				t.Fatalf("failed to accept TCP: %s", err)
+			}
+
+			cl.Conn = conn
+		}
 	})
 
 	t.Run("초과 로그인 시 로그인 실패", func(t *testing.T) {
@@ -153,73 +167,86 @@ func TestWorkerPullOut(t *testing.T) {
 	})
 }
 
-// func TestPlay(t *testing.T) {
-// 	// 유저의 랜덤 위치 이동(스코어가 쌓일 시간을 주기 위해 10초 실행. 복수의 유저가 크래시 없이 game map과 scoreboard에 write하고 점수를 쌓는 것 자체가 테스트의 목적)
-// 	// 정확한 데이터는 예측할 수 없다. 랜덤으로 동전들이 움직일 뿐 더러, 유저의 이동도 랜덤이기 때문에 모두가 0점일 수도 있고 점수가 같거나 높을 수도 있다.
-// 	t.Run("유저의 랜덤 위치 이동", func(t *testing.T) {
-// 		if testing.Short() {
-// 			t.Skip("skipping test in short mode.")
-// 		}
+func TestPlay(t *testing.T) {
+	// 유저의 랜덤 위치 이동(스코어가 쌓일 시간을 주기 위해 10초 실행. 복수의 유저가 크래시 없이 game map과 scoreboard에 write하고 점수를 쌓는 것 자체가 테스트의 목적)
+	// 정확한 데이터는 예측할 수 없다. 랜덤으로 동전들이 움직일 뿐 더러, 유저의 이동도 랜덤이기 때문에 모두가 0점일 수도 있고 점수가 같거나 높을 수도 있다.
+	t.Run("유저의 랜덤 위치 이동", func(t *testing.T) {
+		if testing.Short() {
+			t.Skip("skipping test in short mode.")
+		}
+		var waiter sync.WaitGroup
+		timeout, cancel := context.WithTimeout(context.Background(), time.Second*10)
+		defer cancel()
 
-// 		var allClientWaitGroup sync.WaitGroup
-// 		timeoutReached := false
-// 		overallTimeLimit := 10
+		for _, testClient := range clientListeners {
 
-// 		go func() {
-// 			time.Sleep(time.Second * time.Duration(overallTimeLimit))
-// 			timeoutReached = true
-// 		}()
+			waiter.Add(1)
+			conn, err := net.Dial("tcp", fmt.Sprintf("127.0.0.1:%d", testClient.WorkerPort))
 
-// 		for _, testClient := range clientListeners {
-// 			allClientWaitGroup.Add(1)
-// 			go func() {
-// 				userPosition := &protodef.Position{
-// 					X: 0,
-// 					Y: 0,
-// 				}
+			if err != nil {
+				log.Fatal(err)
+			}
 
-// 				for !timeoutReached {
-// 					// ! 중요 !
-// 					// 원래대로라면 서버가 클라이언트의 위치를 강제로 결정할 수 있다(클라이언트가 유효하지 않은 위치에 존재할 시).
-// 					// 그래서 서버가 보내주는 데이터를 토대로 클라이언트에서 유저의 위치를 재조정하는 등의 동작이 가능하지만럼(멀티 플레이 게임에서 렉 걸릴 때 캐릭터 위치 강제로 결정되는 것 처럼)
-// 					// 이를 구현하려면 클라이언트에 구현되어 있는 것 처럼,
-// 					// 각 테스트 클라이언트가 서버의 TCP 통신을 수신하고 다시 그에 따라 위치를 position을 결정하는
-// 					// 클라이언트의 로직이 서버 테스트에 포함되게 되므로, 서버의 테스트가 아닌 클라이언트의 테스트가 되어버린다.
-// 					// 서버가 애초에 out of bound 혹은 이미 점거된 위치에 대한 판단을 서버가 알아서 진행하므로, 유효하지 않은 위치를 계산하지 않고
-// 					// 클라이언트의 랜덤 포지션을 그대로 전송한다. 이 방법으로도 데이터에 대한 경합성 해결, 요청 무효화 등의 테스트는 가능하다.
-// 					newUserPosition := &protodef.Position{
-// 						X: userPosition.X + game.GenerateRandomDirection(),
-// 						Y: userPosition.Y + game.GenerateRandomDirection(),
-// 					}
+			go func() {
+				userPosition := &protodef.Position{
+					X: 0,
+					Y: 0,
+				}
 
-// 					userStatus := &protodef.Status{
-// 						Id:              testClient.ID,
-// 						CurrentPosition: newUserPosition,
-// 					}
+				for {
+					select {
+					case <-timeout.Done():
+						waiter.Done()
+						return
+					default:
+						// ! 중요 !
+						// 원래대로라면 서버가 클라이언트의 위치를 강제로 결정할 수 있다(클라이언트가 유효하지 않은 위치에 존재할 시).
+						// 그래서 서버가 보내주는 데이터를 토대로 클라이언트에서 유저의 위치를 재조정하는 등의 동작이 가능하지만럼(멀티 플레이 게임에서 렉 걸릴 때 캐릭터 위치 강제로 결정되는 것 처럼)
+						// 이를 구현하려면 클라이언트에 구현되어 있는 것 처럼,
+						// 각 테스트 클라이언트가 서버의 TCP 통신을 수신하고 다시 그에 따라 위치를 position을 결정하는
+						// 클라이언트의 로직이 서버 테스트에 포함되게 되므로, 서버의 테스트가 아닌 클라이언트의 테스트가 되어버린다.
+						// 서버가 애초에 out of bound 혹은 이미 점거된 위치에 대한 판단을 서버가 알아서 진행하므로, 유효하지 않은 위치를 계산하지 않고
+						// 클라이언트의 랜덤 포지션을 그대로 전송한다. 이 방법으로도 데이터에 대한 경합성 해결, 요청 무효화 등의 테스트는 가능하다.
+						newUserPosition := &protodef.Position{
+							X: userPosition.X + game.GenerateRandomDirection(),
+							Y: userPosition.Y + game.GenerateRandomDirection(),
+						}
 
-// 					userPosition = newUserPosition
-// 					marshaledData, err := proto.Marshal(userStatus)
+						if newUserPosition.X < 0 {
+							newUserPosition.X = 0
+						}
+						if newUserPosition.Y < 0 {
+							newUserPosition.Y = 0
+						}
 
-// 					if err != nil {
-// 						panic(err)
-// 					}
-// 					// 유저의 위치를 서버에 전송
-// 					_, err = testClient.Conn.Write(marshaledData)
+						userStatus := &protodef.Status{
+							Id:              testClient.ID,
+							CurrentPosition: newUserPosition,
+						}
 
-// 					if err != nil {
-// 						t.Log(err)
-// 					}
+						userPosition = newUserPosition
+						marshaledData, err := proto.Marshal(userStatus)
 
-// 					time.Sleep(time.Millisecond * 200)
-// 				}
+						if err != nil {
+							panic(err)
+						}
+						// 유저의 위치를 서버에 전송
+						marshaledData = append(marshaledData, '$')
+						_, err = conn.Write(marshaledData)
 
-// 				allClientWaitGroup.Done()
-// 			}()
+						if err != nil {
+							t.Log(err)
+						}
 
-// 			allClientWaitGroup.Wait()
-// 		}
-// 	})
-// }
+						time.Sleep(time.Millisecond * 200)
+					}
+				}
+			}()
+		}
+
+		waiter.Wait()
+	})
+}
 
 func TestWorkerPutBack(t *testing.T) {
 	t.Run("로그아웃(워커 반환)", func(t *testing.T) {
