@@ -1,3 +1,5 @@
 package worker_pool
 
-var workerPool WorkerPool
+var workerPool WorkerPool = WorkerPool{
+	Pool: make(map[string]*Worker),
+}
